% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indiv.R
\name{Indiv}
\alias{Indiv}
\title{R6 Indiv Class}
\description{
R6 class to store individual longitudinal data. It can store unique ID, longitudinal data (time-dependent variables) as well as other time-independent variable.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{UUID}}{Universally Unique Identifier (UUID). This is a mandatory field.}

\item{\code{UUID_Other}}{Sometimes, the UUID is lost and another ID is provided. This field is used to store such information.}

\item{\code{SproutDate}}{Date of sprout. This must be specified is longitudinal data is recorded with un-absolute time scale.}

\item{\code{TimeDependVar}}{Data.frame to store time dependent variable such as phenotype.}

\item{\code{TimeIndpendVar}}{Data.frame to store time independent variable such as cultivation condition.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{TimeIndpendVar}}{Data.frame to store time independent variable such as cultivation condition.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Indiv-new}{\code{Indiv$new()}}
\item \href{#method-Indiv-Plot}{\code{Indiv$Plot()}}
\item \href{#method-Indiv-Summarize}{\code{Indiv$Summarize()}}
\item \href{#method-Indiv-ExtractPheno}{\code{Indiv$ExtractPheno()}}
\item \href{#method-Indiv-UniqueValue}{\code{Indiv$UniqueValue()}}
\item \href{#method-Indiv-clone}{\code{Indiv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Indiv-new"></a>}}
\if{latex}{\out{\hypertarget{method-Indiv-new}{}}}
\subsection{Method \code{new()}}{
Construct Indiv object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Indiv$new(
  RawData,
  ListOfData,
  SproutDate = NULL,
  UUID_Other = NULL,
  duplicate = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RawData}}{one-row Data Frame with several variables as well as UUID}

\item{\code{ListOfData}}{List indicating with column of RawData is corresponding to UUID, TimeDependVar, TimeIndependVar and Date.}

\item{\code{SproutDate}}{Date of Sprout. Not mandatory.}

\item{\code{UUID_Other}}{Data Frame to connect old UUID and new UUID}

\item{\code{duplicate}}{Weather data has duplication or not. This parameter is not used in the function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Indiv-Plot"></a>}}
\if{latex}{\out{\hypertarget{method-Indiv-Plot}{}}}
\subsection{Method \code{Plot()}}{
Method to create plot of individual growth. Parameters are same for plot() function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Indiv$Plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to the base plot function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Indiv-Summarize"></a>}}
\if{latex}{\out{\hypertarget{method-Indiv-Summarize}{}}}
\subsection{Method \code{Summarize()}}{
Method to create and display the summary of specific individual.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Indiv$Summarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Indiv-ExtractPheno"></a>}}
\if{latex}{\out{\hypertarget{method-Indiv-ExtractPheno}{}}}
\subsection{Method \code{ExtractPheno()}}{
Method to display Phenotype of specific individual.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Indiv$ExtractPheno(Target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Target}}{Name of variable to extract.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Indiv-UniqueValue"></a>}}
\if{latex}{\out{\hypertarget{method-Indiv-UniqueValue}{}}}
\subsection{Method \code{UniqueValue()}}{
Method to return unique variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Indiv$UniqueValue()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Indiv-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Indiv-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Indiv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
